Replace link to 404ed/dead pipeline instance links with VSM links
There is no direct replacement for the old Pipeline Instance page, removed many years ago, as this now works via modals on the main GoCD dashboard. The only page which has <pipelineName, pipelineCounter> as args is the VSM page, which is probably a reasonable alternative for now.

Later on a possible alternative would be a route which loads the specific page within pipeline activity corresponding to a given pipeline counter, but that would need GoCD server support.